[FE] refactor: 가드 리팩토링
- 주입 받는 형태로 변경
