+++ client/src/core/Guard/RouteGaurd.tsx
@@ -1,26 +1,20 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import React, { useEffect, useState } from 'react';
 import { RouteProps, useHistory, useParams } from 'react-router-dom';
-import { authHttpClient } from '@api';
 
 interface ParamProps {
   gid: string;
 }
-
-export enum RouteGuardScopes {
-  AUTH = 'AUTH',
-  GROUP_BELONG = 'GROUP_BELONG',
-  GROUP_OWNER = 'GROUP_OWNER',
-}
+export type authenficateReturnType = (gid?: string) => Promise<null>;
 
 /**
  * @param {React.FC} Page 라우트될 페이지 컴포넌트
- * @param {RouteGuardScopes} scope 보호 범위
+ * @param {authenficateReturnType} authenficate 권한 검사 함수
  * @returns 검증 처리 후 컴포넌트
  */
-export default function RouteGuard(
+export function RouteGuard(
   Page: RouteProps['component'],
-  scope?: RouteGuardScopes,
+  authenficate: authenficateReturnType,
 ): (props: RouteProps) => JSX.Element {
   return function guardRoute(props: RouteProps): JSX.Element {
     const Component = Page as React.FC;
@@ -31,21 +25,13 @@ export default function RouteGuard(
     useEffect(() => {
       (async () => {
         try {
-          switch (scope) {
-            case RouteGuardScopes.AUTH:
-              await authHttpClient.isAuthUser();
-              break;
-            case RouteGuardScopes.GROUP_BELONG:
-              await authHttpClient.isGroupBelong(gid);
-              break;
-            case RouteGuardScopes.GROUP_OWNER:
-              await authHttpClient.isGroupOwner(gid);
-              break;
-          }
-          setIsProgress(false);
+          if (!gid) await authenficate();
+          else await authenficate(gid);
         } catch (e: any) {
           history.go(-1);
         }
+
+        setIsProgress(false);
       })();
     }, []);
