+++ client/src/core/Router/routes/AuthGuardRoute.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import { Route, RouteProps } from 'react-router-dom';
-import RouteGuard from '@core/Guard/RouteGaurd';
-import { RouteGuardScopes } from '@core/Guard/RouteGaurd';
+import { RouteGuard } from '@core/Guard';
+import { authHttpClient } from '@api';
 
 function AuthGuardRoute(props: RouteProps): JSX.Element {
   const component: RouteProps['component'] = props.component;
-  return <Route {...props} component={RouteGuard(component, RouteGuardScopes.AUTH)} />;
+  authHttpClient.isAuthUser;
+  return <Route {...props} component={RouteGuard(component, authHttpClient.isAuthUser)} />;
 }
 
 export default AuthGuardRoute;
