+++ client/src/core/Router/routes/GroupBelongGuardRoute.tsx
@@ -1,11 +1,16 @@
 import React from 'react';
 import { Route, RouteProps } from 'react-router-dom';
-import RouteGuard from '@core/Guard/RouteGaurd';
-import { RouteGuardScopes } from '@core/Guard/RouteGaurd';
+import { RouteGuard, authenficateReturnType } from '@core/Guard';
+import { authHttpClient } from '@api';
 
 function GroupBelongGuardRoute(props: RouteProps): JSX.Element {
   const component: RouteProps['component'] = props.component;
-  return <Route {...props} component={RouteGuard(component, RouteGuardScopes.GROUP_BELONG)} />;
+  return (
+    <Route
+      {...props}
+      component={RouteGuard(component, authHttpClient.isGroupBelong as authenficateReturnType)}
+    />
+  );
 }
 
 export default GroupBelongGuardRoute;
