+++ client/src/core/Router/routes/GroupOwnerGuardRoute.tsx
@@ -1,11 +1,16 @@
 import React from 'react';
 import { Route, RouteProps } from 'react-router-dom';
-import RouteGuard from '@core/Guard/RouteGaurd';
-import { RouteGuardScopes } from '@core/Guard/RouteGaurd';
+import { RouteGuard, authenficateReturnType } from '@core/Guard';
+import { authHttpClient } from '@api';
 
 function GroupOwnerGuardRoute(props: RouteProps): JSX.Element {
   const component: RouteProps['component'] = props.component;
-  return <Route {...props} component={RouteGuard(component, RouteGuardScopes.GROUP_OWNER)} />;
+  return (
+    <Route
+      {...props}
+      component={RouteGuard(component, authHttpClient.isGroupOwner as authenficateReturnType)}
+    />
+  );
 }
 
 export default GroupOwnerGuardRoute;
